+++ gazebo/gui/GLWidget.cc
@@ -67,6 +67,10 @@ GLWidget::GLWidget(QWidget *_parent)
   this->dataPtr->copyEntityName = "";
   this->dataPtr->modelEditorEnabled = false;
 
+  this->dataPtr->updateTimer = new QTimer(this);
+  connect(this->dataPtr->updateTimer, SIGNAL(timeout()),
+          this, SLOT(update()));
+
   this->setFocusPolicy(Qt::StrongFocus);
 
   this->dataPtr->windowId = -1;
@@ -285,8 +289,6 @@ void GLWidget::paintEvent(QPaintEvent *_e)
   {
     event::Events::preRender();
   }
-
-  this->update();
 
   _e->accept();
 }
@@ -875,6 +877,11 @@ void GLWidget::ViewScene(rendering::ScenePtr _scene)
   double pitch = atan2(-delta.z, sqrt(delta.x*delta.x + delta.y*delta.y));
   this->dataPtr->userCamera->SetDefaultPose(math::Pose(camPos,
         math::Vector3(0, pitch, yaw)));
+
+  // Update at the camera's update rate
+  this->dataPtr->updateTimer->start(
+      static_cast<int>(
+        std::round(1000.0 / (4*this->dataPtr->userCamera->RenderRate()))));
 }
 
 /////////////////////////////////////////////////