+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## Gazebo 7.x.x (2016-xx-xx)
 
+1. Update QT render loop to throttle based on UserCamera::RenderRate. 
+    * [Pull request 2476](https://bitbucket.org/osrf/gazebo/pull-request/2476)
+    * [Issue 1560](https://bitbucket.org/osrf/gazebo/issues/1560)
+
 1. Generate visualization on demand, instead of on load. This helps to
    reduce load time.
     * [Pull request 2457](https://bitbucket.org/osrf/gazebo/pull-request/2457)